+++ crates/primitives/transactions/src/to_starknet_core_transaction.rs
@@ -4,8 +4,6 @@ use mp_felt::Felt252Wrapper;
 use mp_hashers::HasherT;
 use starknet_crypto::FieldElement;
 
-use super::compute_hash::ComputeTransactionHash;
-
 fn cast_vec_of_felt_252_wrappers(data: Vec<Felt252Wrapper>) -> Vec<FieldElement> {
     // Non-copy but less dangerous than transmute
     // https://doc.rust-lang.org/std/mem/fn.transmute.html#alternatives
@@ -15,12 +13,10 @@ fn cast_vec_of_felt_252_wrappers(data: Vec<Felt252Wrapper>) -> Vec<FieldElement>
 
 pub fn to_starknet_core_tx<H: HasherT>(
     tx: super::Transaction,
-    chain_id: Felt252Wrapper,
+    transaction_hash: FieldElement,
 ) -> starknet_core::types::Transaction {
     match tx {
         super::Transaction::Declare(tx) => {
-            let tx_hash = tx.compute_hash::<H>(chain_id, false);
-
             let tx = match tx {
                 super::DeclareTransaction::V0(super::DeclareTransactionV0 {
                     max_fee,
@@ -29,7 +25,7 @@ pub fn to_starknet_core_tx<H: HasherT>(
                     class_hash,
                     sender_address,
                 }) => starknet_core::types::DeclareTransaction::V0(starknet_core::types::DeclareTransactionV0 {
-                    transaction_hash: tx_hash.0,
+                    transaction_hash,
                     max_fee: max_fee.into(),
                     signature: cast_vec_of_felt_252_wrappers(signature),
                     class_hash: class_hash.into(),
@@ -42,7 +38,7 @@ pub fn to_starknet_core_tx<H: HasherT>(
                     class_hash,
                     sender_address,
                 }) => starknet_core::types::DeclareTransaction::V1(starknet_core::types::DeclareTransactionV1 {
-                    transaction_hash: tx_hash.0,
+                    transaction_hash,
                     max_fee: max_fee.into(),
                     signature: cast_vec_of_felt_252_wrappers(signature),
                     nonce: nonce.into(),
@@ -57,7 +53,7 @@ pub fn to_starknet_core_tx<H: HasherT>(
                     sender_address,
                     compiled_class_hash,
                 }) => starknet_core::types::DeclareTransaction::V2(starknet_core::types::DeclareTransactionV2 {
-                    transaction_hash: tx_hash.0,
+                    transaction_hash,
                     max_fee: max_fee.into(),
                     signature: cast_vec_of_felt_252_wrappers(signature),
                     nonce: nonce.into(),
@@ -70,10 +66,8 @@ pub fn to_starknet_core_tx<H: HasherT>(
             starknet_core::types::Transaction::Declare(tx)
         }
         super::Transaction::DeployAccount(tx) => {
-            let tx_hash = tx.compute_hash::<H>(chain_id, false);
-
             let tx = starknet_core::types::DeployAccountTransaction {
-                transaction_hash: tx_hash.0,
+                transaction_hash,
                 max_fee: tx.max_fee.into(),
                 signature: cast_vec_of_felt_252_wrappers(tx.signature),
                 nonce: tx.nonce.into(),
@@ -85,18 +79,15 @@ pub fn to_starknet_core_tx<H: HasherT>(
             starknet_core::types::Transaction::DeployAccount(tx)
         }
         super::Transaction::Invoke(tx) => {
-            let tx_hash = tx.compute_hash::<H>(chain_id, false);
-
             let tx = match tx {
                 super::InvokeTransaction::V0(super::InvokeTransactionV0 {
                     max_fee,
                     signature,
-                    nonce: _,
                     contract_address,
                     entry_point_selector,
                     calldata,
                 }) => starknet_core::types::InvokeTransaction::V0(starknet_core::types::InvokeTransactionV0 {
-                    transaction_hash: tx_hash.0,
+                    transaction_hash,
                     max_fee: max_fee.into(),
                     signature: cast_vec_of_felt_252_wrappers(signature),
                     contract_address: contract_address.into(),
@@ -110,7 +101,7 @@ pub fn to_starknet_core_tx<H: HasherT>(
                     sender_address,
                     calldata,
                 }) => starknet_core::types::InvokeTransaction::V1(starknet_core::types::InvokeTransactionV1 {
-                    transaction_hash: tx_hash.0,
+                    transaction_hash,
                     max_fee: max_fee.into(),
                     signature: cast_vec_of_felt_252_wrappers(signature),
                     nonce: nonce.into(),
@@ -122,10 +113,8 @@ pub fn to_starknet_core_tx<H: HasherT>(
             starknet_core::types::Transaction::Invoke(tx)
         }
         super::Transaction::L1Handler(tx) => {
-            let tx_hash = tx.compute_hash::<H>(chain_id, false);
-
             let tx = starknet_core::types::L1HandlerTransaction {
-                transaction_hash: tx_hash.0,
+                transaction_hash,
                 version: 0,
                 nonce: tx.nonce,
                 contract_address: tx.contract_address.into(),